feat(web): allow a player to manage their balance
