+++ web/src/routes/_layout/account.tsx
@@ -1,30 +1,14 @@
-import { ChangeEvent, useState } from "react";
-import { useForm } from "react-hook-form";
+import { useState } from "react";
 
-import { zodResolver } from "@hookform/resolvers/zod";
-import {
-  QueryKey,
-  queryOptions,
-  useMutation,
-  useQueryClient,
-  useSuspenseQuery,
-} from "@tanstack/react-query";
+import { queryOptions, useSuspenseQuery } from "@tanstack/react-query";
 import { createFileRoute, Link, redirect } from "@tanstack/react-router";
-import { format } from "date-fns";
-import {
-  CalendarIcon,
-  Download,
-  Gamepad2,
-  LoaderCircle,
-  LogOut,
-  Upload,
-  User as UserIcon,
-} from "lucide-react";
-import { z } from "zod";
+import { Download, Gamepad2, UserIcon } from "lucide-react";
 
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
+import { AccountDetails } from "@/components/account/account-details";
+import { AddFunds } from "@/components/account/add-funds";
+import { SignOut } from "@/components/account/sign-out";
+import { UserAvatar } from "@/components/account/user-avatar";
 import { Button } from "@/components/ui/button";
-import { Calendar } from "@/components/ui/calendar";
 import {
   Card,
   CardContent,
@@ -32,48 +16,12 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { User } from "@/domain/user";
-import { toast, useToast } from "@/hooks/use-toast";
-import { getUser, updateUser, uploadMultimedia } from "@/lib/api";
-import { clearToken, decodeTokenPayload, getToken } from "@/lib/auth";
-import { COUNTRIES, COUNTRIES_MAP } from "@/lib/constants";
-import { Conflict, ContentTooLarge } from "@/lib/errors";
-import { cn, formatCurrency, getInitials } from "@/lib/utils";
-import { accountDetailsSchema } from "@/lib/zod";
-
-const userQueryKey: QueryKey = ["user"];
+import { getUser } from "@/lib/api";
+import { decodeTokenPayload, getToken } from "@/lib/auth";
+import { COUNTRIES_MAP } from "@/lib/constants";
+import { userQueryKey } from "@/lib/query-keys";
+import { formatCurrency } from "@/lib/utils";
 
 /**
  * Query options for retrieving the signed in user.
@@ -137,7 +85,7 @@ function Component() {
             <p className="text-2xl font-bold">{formatCurrency(user.balance)}</p>
             <div className="flex items-center gap-2 mt-2">
               <SignOut />
-              <AddFunds />
+              <AddFunds balance={user.balance} id={user.id} />
             </div>
           </div>
         </CardHeader>
@@ -198,445 +146,5 @@ function Component() {
         </CardContent>
       </Card>
     </div>
-  );
-}
-
-function AddFunds() {
-  return (
-    <Dialog>
-      <DialogTrigger asChild>
-        <Button>Add Funds</Button>
-      </DialogTrigger>
-      <DialogContent className="sm:max-w-[425px]">
-        <DialogHeader>
-          <DialogTitle>Add Funds to Your Account</DialogTitle>
-          <DialogDescription>
-            Enter the amount you'd like to add to your account balance.
-          </DialogDescription>
-        </DialogHeader>
-        <form>
-          <div className="grid gap-4 py-4">
-            <div className="grid grid-cols-4 items-center gap-4">
-              <Label className="text-right" htmlFor="amount">
-                Amount
-              </Label>
-              <Input
-                required
-                className="col-span-3"
-                id="amount"
-                min="0.01"
-                placeholder="0.00"
-                step="0.01"
-                type="number"
-              />
-            </div>
-          </div>
-          <DialogFooter>
-            <Button type="submit">Add Funds</Button>
-          </DialogFooter>
-        </form>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-function UserAvatar(props: { id: string; displayName: string; url?: string }) {
-  const queryClient = useQueryClient();
-  const mutation = useMutation({
-    async mutationFn(file: File) {
-      const multimedia = await uploadMultimedia(file);
-
-      await updateUser(props.id, { pictureMultimediaId: multimedia.id });
-    },
-    async onSuccess() {
-      await queryClient.invalidateQueries({ queryKey: userQueryKey });
-    },
-    onError(error) {
-      if (error instanceof ContentTooLarge) {
-        toast({
-          variant: "destructive",
-          title: "Picture size must be smaller than 2MB",
-        });
-        return;
-      }
-
-      toast({
-        variant: "destructive",
-        title: "Oops! An unexpected error occurred",
-        description: "Please try again later or contact the support team.",
-      });
-    },
-  });
-
-  /**
-   * Handles file upload.
-   * @param event Input change event.
-   */
-  function handleFileUpload(event: ChangeEvent<HTMLInputElement>) {
-    const files = event.target.files;
-    if (!files) {
-      return;
-    }
-
-    mutation.mutate(files[0]);
-  }
-
-  return (
-    <Avatar asChild className="relative size-24 group cursor-pointer">
-      <label>
-        <AvatarImage
-          alt="Gamer Avatar"
-          className="object-cover"
-          src={props.url}
-        />
-        <AvatarFallback>{getInitials(props.displayName)}</AvatarFallback>
-        {mutation.isPending ? (
-          <>
-            <div className="absolute size-full bg-black opacity-70" />
-            <div className="absolute left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 ">
-              <LoaderCircle className="animate-spin" />
-            </div>
-          </>
-        ) : (
-          <>
-            <div className="absolute size-full bg-black opacity-0 group-hover:opacity-70 transition-opacity" />
-            <Upload className="absolute left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 opacity-0 group-hover:opacity-100 transition-opacity" />
-          </>
-        )}
-        <Input
-          accept="image/png, image/jpeg"
-          className="hidden"
-          type="file"
-          onChange={handleFileUpload}
-        />
-      </label>
-    </Avatar>
-  );
-}
-
-function SignOut() {
-  /**
-   * Signs out a user and reloads the current page.
-   */
-  function handleClick() {
-    clearToken();
-    window.location.reload();
-  }
-
-  return (
-    <Button variant="ghost" onClick={handleClick}>
-      <LogOut /> Sign Out
-    </Button>
-  );
-}
-
-function AccountDetails(props: { user: User; country: string }) {
-  const [isEditMode, setEditMode] = useState(false);
-
-  if (isEditMode) {
-    return (
-      <EditAccountDetails
-        user={props.user}
-        onCancel={() => setEditMode(false)}
-        onSave={() => setEditMode(false)}
-      />
-    );
-  }
-
-  return (
-    <ViewAccountDetails
-      country={props.country}
-      user={props.user}
-      onEdit={() => setEditMode(true)}
-    />
-  );
-}
-
-function ViewAccountDetails(props: {
-  user: User;
-  country: string;
-  onEdit: () => void;
-}) {
-  return (
-    <div className="space-y-4">
-      <div className="flex h-10 justify-between items-center mb-4">
-        <h3 className="text-lg font-semibold">Account Details</h3>
-        <Button variant="secondary" onClick={props.onEdit}>
-          Edit Profile
-        </Button>
-      </div>
-      <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-        <div>
-          <p className="text-sm font-medium text-muted-foreground">Username</p>
-          <p className="text-lg">{props.user.username}</p>
-        </div>
-        <div>
-          <p className="text-sm font-medium text-muted-foreground">Email</p>
-          <p className="text-lg">{props.user.email}</p>
-        </div>
-        <div>
-          <p className="text-sm font-medium text-muted-foreground">Full Name</p>
-          <p className="text-lg">{props.user.displayName}</p>
-        </div>
-        <div>
-          <p className="text-sm font-medium text-muted-foreground">
-            Date of Birth
-          </p>
-          <p className="text-lg">
-            {format(props.user.dateOfBirth, "dd/MM/yyyy")}
-          </p>
-        </div>
-        <div>
-          <p className="text-sm font-medium text-muted-foreground">Country</p>
-          <p className="text-lg">{props.country}</p>
-        </div>
-        <div>
-          <p className="text-sm font-medium text-muted-foreground">VAT</p>
-          <p className="text-lg">{props.user.vatin}</p>
-        </div>
-      </div>
-      <div>
-        <p className="text-sm font-medium text-muted-foreground">Address</p>
-        <p className="text-lg">{props.user.address}</p>
-      </div>
-    </div>
-  );
-}
-
-type AccountDetailsSchemaType = z.infer<typeof accountDetailsSchema>;
-
-function EditAccountDetails(props: {
-  user: User;
-  onCancel: () => void;
-  onSave: () => void;
-}) {
-  const queryClient = useQueryClient();
-  const form = useForm<AccountDetailsSchemaType>({
-    resolver: zodResolver(accountDetailsSchema),
-    defaultValues: {
-      username: props.user.username,
-      email: props.user.email,
-      displayName: props.user.displayName,
-      dateOfBirth: new Date(props.user.dateOfBirth),
-      country: props.user.country.toUpperCase(),
-      address: props.user.address,
-      vatin: props.user.vatin,
-    },
-  });
-  const { toast } = useToast();
-  const mutation = useMutation({
-    async mutationFn(data: AccountDetailsSchemaType) {
-      await updateUser(props.user.id, {
-        username: data.username,
-        email: data.email,
-        displayName: data.displayName,
-        dateOfBirth: format(data.dateOfBirth, "yyyy-MM-dd"),
-        country: data.country.toUpperCase(),
-        address: data.address,
-        vatin: data.vatin,
-      });
-    },
-    async onSuccess() {
-      await queryClient.invalidateQueries({ queryKey: userQueryKey });
-      props.onSave();
-    },
-    onError(error) {
-      if (error instanceof Conflict) {
-        switch (error.code) {
-          case "user_username_already_exists":
-            form.setError("username", { message: "Username already exists" });
-            break;
-
-          case "user_email_already_exists":
-            form.setError("email", { message: "Email already exists" });
-            break;
-
-          case "user_vatin_already_exists":
-            form.setError("vatin", { message: "VAT already exists" });
-            break;
-        }
-        return;
-      }
-
-      toast({
-        variant: "destructive",
-        title: "Oops! An unexpected error occurred",
-        description: "Please try again later or contact the support team.",
-      });
-    },
-  });
-
-  /**
-   * Handles form submission.
-   * @param data Form data.
-   */
-  function onSubmit(data: AccountDetailsSchemaType) {
-    mutation.mutate(data);
-  }
-
-  return (
-    <>
-      <div className="flex items-center h-10 mb-4">
-        <h3 className="text-lg font-semibold">Account Details</h3>
-      </div>
-      <Form {...form}>
-        <form className="space-y-4" onSubmit={form.handleSubmit(onSubmit)}>
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-            <FormField
-              control={form.control}
-              name="username"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Username</FormLabel>
-                  <FormControl>
-                    <Input placeholder="Enter your username" {...field} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <FormField
-              control={form.control}
-              name="email"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Email</FormLabel>
-                  <FormControl>
-                    <Input
-                      placeholder="Enter your email"
-                      type="email"
-                      {...field}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <FormField
-              control={form.control}
-              name="displayName"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Full Name</FormLabel>
-                  <FormControl>
-                    <Input placeholder="Enter your full name" {...field} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <FormField
-              control={form.control}
-              name="dateOfBirth"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Date of Birth</FormLabel>
-                  <Popover>
-                    <PopoverTrigger asChild>
-                      <FormControl>
-                        <Button
-                          variant="outline"
-                          className={cn(
-                            "w-full pl-3 text-left font-normal",
-                            !field.value && "text-muted-foreground",
-                          )}
-                        >
-                          {field.value ? (
-                            format(field.value, "dd/MM/yyyy")
-                          ) : (
-                            <span>Enter your date of birth</span>
-                          )}
-                          <CalendarIcon className="ml-auto h-4 w-4 opacity-50" />
-                        </Button>
-                      </FormControl>
-                    </PopoverTrigger>
-                    <PopoverContent align="start" className="w-auto p-0">
-                      <Calendar
-                        initialFocus
-                        disabled={(date) => date > new Date()}
-                        mode="single"
-                        selected={field.value}
-                        onSelect={field.onChange}
-                      />
-                    </PopoverContent>
-                  </Popover>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <FormField
-              control={form.control}
-              name="country"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Country</FormLabel>
-                  <Select
-                    defaultValue={field.value}
-                    onValueChange={field.onChange}
-                  >
-                    <FormControl>
-                      <SelectTrigger className="border-input">
-                        <SelectValue placeholder="Select a country" />
-                      </SelectTrigger>
-                    </FormControl>
-                    <SelectContent>
-                      {COUNTRIES.map((country) => {
-                        return (
-                          <SelectItem key={country.code} value={country.code}>
-                            {country.name}
-                          </SelectItem>
-                        );
-                      })}
-                    </SelectContent>
-                  </Select>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <FormField
-              control={form.control}
-              name="vatin"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>VAT</FormLabel>
-                  <FormControl>
-                    <Input placeholder="Enter your VAT" {...field} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-          </div>
-
-          <FormField
-            control={form.control}
-            name="address"
-            render={({ field }) => (
-              <FormItem>
-                <FormLabel>Address</FormLabel>
-                <FormControl>
-                  <Input placeholder="Enter your address" {...field} />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-
-          <div className="flex w-full justify-end items-center gap-2 mt-4">
-            <Button variant="ghost" onClick={props.onCancel}>
-              Cancel
-            </Button>
-            <Button disabled={mutation.isPending} type="submit">
-              Save
-            </Button>
-          </div>
-        </form>
-      </Form>
-    </>
   );
 }
